Power Analysis for Rates using Bayesian Model

Uses rate_sim() and rate_effect() to simulate the statistical power (probability of a p-value < 0.05). If p2 is NULL estimates the power for detecting whether p1 is different from 0.5. Otherwise estimates the power for detecting a difference between the two groups.

rate2_power_analysis_bayesian(
  p1,
  n,
  p2 = NULL,
  alpha1 = 1,
  beta1 = 1,
  alpha2 = alpha1,
  beta2 = beta1,
  alternative = "two.sided",
  nsims = 1000,
  niter = 10000
)

Arguments

p1

A numeric scalar of the proportion of individuals expected to return from the first group.

n

The total number of individuals in the study (divided evenly among groups) or if a numeric vector the number individuals in each group.

p2

A numeric scalar of the proportion of individuals expected to return from the second group.

alpha1

A numeric vector of the first shape parameter for the beta distribution for the first group.

beta1

A numeric vector of the second shape parameter for the beta distribution for the first group.

alpha2

A numeric vector of the first shape parameter for the beta distribution for the second group.

beta2

A numeric vector of the second shape parameter for the beta distribution for the second group.

alternative

A string specifying whether to perform a "two.sided", or "less" or "greater" one-sided test.

nsims

Number of simulations to perform.

niter

The number of iterations per chain for MCMC sampling.

Value

A real scalar of the power.

See also

Examples

rate2_power_analysis_bayesian(0.9, 10, nsims = 100)
#> Registered S3 method overwritten by 'mcmcr':
#>   method               from 
#>   as.mcmc.list.mcarray rjags
#> [1] 0.68
rate2_power_analysis_bayesian(0.005, p2 = 0.0075, n = 1000, alternative = "greater", nsims = 100)
#> [1] 0.13