Estimate Rates with Credible Limits and P-Values

Estimates rates with 95% credible limits. The credible limits and p-values are estimate using MCMC sample. The p-value for the first group is for p1 = 0.5 while the p-value for subsequent groups is for pn = p1.

rate_effect_bayesian(
  r,
  n,
  alpha = 1,
  beta = 1,
  alternative = "two.sided",
  niter = 10000
)

Arguments

r

A numeric vector of the number of individuals returning in each group.

n

The total number of individuals in the study (divided evenly among groups) or if a numeric vector the number individuals in each group.

alpha

A numeric vector of the first shape parameter for the beta distribution.

beta

A numeric vector of the second shape parameter for the beta distribution.

alternative

A string specifying whether to perform a "two.sided", or "less" or "greater" one-sided test.

niter

The number of iterations per chain for MCMC sampling.

Value

A data frame with eight columns of group, r, n, alpha, beta, the estimate of the rate, the lower and upper 95% confidence limits and the p-value.

See also

Examples

set.seed(99)
rate_effect_bayesian(0, 5)
#>   group r n alpha beta estimate       lower     upper     pvalue
#> 1     1 0 5     1    1 0.109223 0.003733137 0.4658722 0.03343222
rate_effect_bayesian(0, 5, 10, 10)
#>   group r n alpha beta  estimate     lower     upper    pvalue
#> 1     1 0 5    10   10 0.3966767 0.2225761 0.5936213 0.3078231
rate_effect_bayesian(c(0,5), c(5,5), 10, 10)
#>   group r n alpha beta  estimate     lower     upper    pvalue
#> 1     1 0 5    10   10 0.3992186 0.2202440 0.5974719 0.3060231
#> 2     2 5 5    10   10 0.6019532 0.4062801 0.7776257 0.1530282